Setup Calendar Component/#63 Creating the calendar component
